refactor(editor): extract summaries panel
Mechanical module split: summary tile toggles incl. maintenance-users,
hidden-cameras and security-extra-entities pickers move to
panels/SummariesPanel.ts. Moved-line hardening: stateFor() for the
camera friendly-name lookup. Also drops an orphaned comment block the
weather-sensors move left behind.
