+++ src/editor/StrategyEditor.ts
@@ -38,12 +38,7 @@ import {
   renderCustomSectionsSection,
 } from './panels/CustomConfigPanels';
 import { renderSectionOrderPanel } from './panels/SectionOrderPanel';
-import {
-  getAllEntitiesForSelect,
-  getFilteredEntities,
-} from './entity-options';
-import { Registry } from '../Registry';
-import { collectCameraBlocks } from '../views/CctvViewStrategy';
+import { renderSummariesSection } from './panels/SummariesPanel';
 import { findUpsEntityGroups } from '../views/RoomViewStrategy';
 import { isBadgeCandidate, isDefaultShowName, resolveShowName } from '../utils/badge-utils';
 import { mergeStacksOrder } from '../utils/name-utils';
@@ -124,7 +119,7 @@ class Simon42DashboardStrategyEditor extends LitElement implements StrategyEdito
     return html`
       <div class="card-config">
         ${renderOverviewSection(this)}
-        ${this._renderSummariesSection()}
+        ${renderSummariesSection(this)}
         ${renderFavoritesSection(this)}
         ${renderLightFavoritesSection(this)}
 
@@ -415,403 +410,6 @@ class Simon42DashboardStrategyEditor extends LitElement implements StrategyEdito
   };
 
   // -- Overview section --------------------------------------------------
-
-  private _renderSummariesSection(): TemplateResult {
-    const summariesColumns = this._config.summaries_columns || 2;
-    const showLightSummary = this._config.show_light_summary !== false;
-    const groupLightsByFloors = this._config.group_lights_by_floors === true;
-    const nestedLightGroups = this._config.nested_light_groups === true;
-    const showCoversSummary = this._config.show_covers_summary !== false;
-    const showPartiallyOpenCovers = this._config.show_partially_open_covers === true;
-    const groupCoversByFloors = this._config.group_covers_by_floors === true;
-    const showSecuritySummary = this._config.show_security_summary !== false;
-    const showClimateSummary = this._config.show_climate_summary === true;
-    const showBatterySummary = this._config.show_battery_summary !== false;
-    const hideMobileAppBatteries = this._config.hide_mobile_app_batteries === true;
-    const hideBatteryNotesEntities = this._config.hide_battery_notes_entities === true;
-    const batteryCriticalThreshold = this._config.battery_critical_threshold ?? 20;
-    const batteryLowThreshold = this._config.battery_low_threshold ?? 50;
-    const showAreaInBatteryView = this._config.show_area_in_battery_view === true;
-    const unavailableBatteriesBucket = this._config.unavailable_batteries_bucket === 'critical' ? 'critical' : 'good';
-
-    return html`
-      <div class="section">
-        <div class="section-title">${localize('editor.section_summaries')}</div>
-
-        <div class="form-row">
-          <input type="radio" id="summaries-2-columns" name="summaries-columns" value="2"
-            ?checked=${summariesColumns === 2}
-            @change=${() => this._summariesColumnsChanged(2)} />
-          <label for="summaries-2-columns">${localize('editor.columns_2')}</label>
-        </div>
-        <div class="form-row">
-          <input type="radio" id="summaries-4-columns" name="summaries-columns" value="4"
-            ?checked=${summariesColumns === 4}
-            @change=${() => this._summariesColumnsChanged(4)} />
-          <label for="summaries-4-columns">${localize('editor.columns_4')}</label>
-        </div>
-        <div class="description">${localize('editor.columns_desc')}</div>
-
-        ${this._renderCheckbox('show-light-summary', localize('editor.show_light_summary'), showLightSummary,
-          (checked) => this._toggleChanged('show_light_summary', checked, true))}
-
-        ${this._renderCheckbox('group-lights-by-floors', localize('editor.group_lights_by_floors'), groupLightsByFloors,
-          (checked) => this._toggleChanged('group_lights_by_floors', checked, false))}
-        <div class="description">${localize('editor.group_lights_by_floors_desc')}</div>
-
-        ${this._renderCheckbox('lights-sort-by-name', localize('editor.lights_sort_by_name'), this._config.lights_sort_by === 'name',
-          (checked) => this._lightsSortByChanged(checked))}
-        <div class="description">${localize('editor.lights_sort_by_name_desc')}</div>
-
-        ${this._renderCheckbox('nested-light-groups', localize('editor.nested_light_groups'), nestedLightGroups,
-          (checked) => this._toggleChanged('nested_light_groups', checked, false))}
-        <div class="description">${localize('editor.nested_light_groups_desc')}</div>
-
-        ${this._renderCheckbox('show-covers-summary', localize('editor.show_covers_summary'), showCoversSummary,
-          (checked) => this._toggleChanged('show_covers_summary', checked, true))}
-
-        <div style="margin-left: 26px; margin-bottom: 8px;">
-          ${this._renderCheckbox('show-partially-open-covers', localize('editor.show_partially_open_covers'), showPartiallyOpenCovers,
-            (checked) => this._toggleChanged('show_partially_open_covers', checked, false))}
-          <div class="description">${localize('editor.show_partially_open_covers_desc')}</div>
-
-          ${this._renderCheckbox('group-covers-by-floors', localize('editor.group_covers_by_floors'), groupCoversByFloors,
-            (checked) => this._toggleChanged('group_covers_by_floors', checked, false))}
-          <div class="description">${localize('editor.group_covers_by_floors_desc')}</div>
-        </div>
-
-        ${this._renderCheckbox('show-security-summary', localize('editor.show_security_summary'), showSecuritySummary,
-          (checked) => this._toggleChanged('show_security_summary', checked, true))}
-
-        <div style="margin-left: 26px; margin-bottom: 8px;">
-          ${this._renderCheckbox('show-cameras-in-security', localize('editor.show_cameras_in_security'), this._config.show_cameras_in_security === true,
-            (checked) => this._toggleChanged('show_cameras_in_security', checked, false))}
-          <div class="description">${localize('editor.show_cameras_in_security_desc')}</div>
-
-          ${this._config.show_cameras_in_security === true ? this._renderHiddenCamerasPicker() : nothing}
-
-          ${this._renderCheckbox('group-security-by-areas', localize('editor.group_security_by_areas'), this._config.group_security_by_areas === true,
-            (checked) => this._toggleChanged('group_security_by_areas', checked, false))}
-          <div class="description">${localize('editor.group_security_by_areas_desc')}</div>
-
-          ${this._renderCheckbox('show-security-activity', localize('editor.show_security_activity'), this._config.show_security_activity !== false,
-            (checked) => this._toggleChanged('show_security_activity', checked, true))}
-          <div class="description">${localize('editor.show_security_activity_desc')}</div>
-
-          ${this._config.show_security_activity !== false && this._config.group_security_by_areas !== true ? html`
-            <div style="margin-left: 26px;">
-              ${this._renderCheckbox('security-activity-at-end', localize('editor.security_activity_at_end'), this._config.security_activity_position === 'end',
-                (checked) => this._securityActivityPositionChanged(checked))}
-            </div>
-          ` : nothing}
-
-          ${this._renderSecurityExtraEntitiesPicker()}
-        </div>
-
-        ${this._renderCheckbox('show-climate-summary', localize('editor.show_climate_summary'), showClimateSummary,
-          (checked) => this._toggleChanged('show_climate_summary', checked, false))}
-        <div class="description">${localize('editor.show_climate_summary_desc')}</div>
-
-        ${this._renderCheckbox('show-camera-view', localize('editor.show_camera_view'), this._config.show_camera_view === true,
-          (checked) => this._toggleChanged('show_camera_view', checked, false))}
-        <div class="description">${localize('editor.show_camera_view_desc')}</div>
-
-        <div style="margin-left: 26px; margin-bottom: 8px;">
-          ${this._renderCheckbox('show-camera-events', localize('editor.show_camera_events'), this._config.show_camera_events === true,
-            (checked) => this._toggleChanged('show_camera_events', checked, false))}
-          <div class="description">${localize('editor.show_camera_events_desc')}</div>
-
-          ${this._config.show_camera_view === true && this._config.show_cameras_in_security !== true
-            ? this._renderHiddenCamerasPicker()
-            : nothing}
-        </div>
-
-        ${this._renderCheckbox('show-battery-summary', localize('editor.show_battery_summary'), showBatterySummary,
-          (checked) => this._toggleChanged('show_battery_summary', checked, true))}
-
-        <div style="margin-left: 26px; margin-bottom: 8px;">
-          ${this._renderCheckbox('hide-mobile-app-batteries', localize('editor.hide_mobile_app_batteries'), hideMobileAppBatteries,
-            (checked) => this._toggleChanged('hide_mobile_app_batteries', checked, false))}
-          <div class="description">${localize('editor.hide_mobile_app_batteries_desc')}</div>
-
-          ${this._renderCheckbox('show-battery-view', localize('editor.show_battery_view'), this._config.show_battery_view === true,
-            (checked) => this._toggleChanged('show_battery_view', checked, false))}
-          <div class="description">${localize('editor.show_battery_view_desc')}</div>
-
-          ${this._renderCheckbox('show-area-in-battery-view', localize('editor.show_area_in_battery_view'), showAreaInBatteryView,
-            (checked) => this._toggleChanged('show_area_in_battery_view', checked, false))}
-          <div class="description">${localize('editor.show_area_in_battery_view_desc')}</div>
-          ${this._renderCheckbox('hide-battery-notes-entities', localize('editor.hide_battery_notes_entities'), hideBatteryNotesEntities,
-            (checked) => this._toggleChanged('hide_battery_notes_entities', checked, false))}
-          <div class="description">${localize('editor.hide_battery_notes_entities_desc')}</div>
-
-          <div style="font-size: 13px; font-weight: 500; color: var(--primary-text-color); margin-top: 12px; margin-bottom: 4px;">
-            ${localize('editor.battery_thresholds')}
-          </div>
-          <div class="form-row">
-            <label for="battery-critical-threshold" style="min-width: 140px;">${localize('editor.battery_critical_below')}</label>
-            <input type="number" id="battery-critical-threshold" min="1" max="99"
-              .value=${String(batteryCriticalThreshold)}
-              style="width: 70px;"
-              @change=${this._batteryCriticalChanged} /> %
-          </div>
-          <div class="form-row">
-            <label for="battery-low-threshold" style="min-width: 140px;">${localize('editor.battery_low_below')}</label>
-            <input type="number" id="battery-low-threshold" min="1" max="99"
-              .value=${String(batteryLowThreshold)}
-              style="width: 70px;"
-              @change=${this._batteryLowChanged} /> %
-          </div>
-          <div class="description">${localize('editor.battery_thresholds_desc')}</div>
-
-          <div style="font-size: 13px; font-weight: 500; color: var(--primary-text-color); margin-top: 12px; margin-bottom: 4px;">
-            ${localize('editor.unavailable_batteries_bucket')}
-          </div>
-          <div class="form-row">
-            <input type="radio" id="unavailable-batteries-critical" name="unavailable-batteries-bucket" value="critical"
-              ?checked=${unavailableBatteriesBucket === 'critical'}
-              @change=${() => this._unavailableBatteriesBucketChanged('critical')} />
-            <label for="unavailable-batteries-critical">${localize('editor.unavailable_batteries_critical')}</label>
-          </div>
-          <div class="form-row">
-            <input type="radio" id="unavailable-batteries-good" name="unavailable-batteries-bucket" value="good"
-              ?checked=${unavailableBatteriesBucket === 'good'}
-              @change=${() => this._unavailableBatteriesBucketChanged('good')} />
-            <label for="unavailable-batteries-good">${localize('editor.unavailable_batteries_good')}</label>
-          </div>
-          <div class="description">${localize('editor.unavailable_batteries_bucket_desc')}</div>
-        </div>
-
-        ${this._renderCheckbox('show-maintenance-summary', localize('editor.show_maintenance_summary'), this._config.show_maintenance_summary === true,
-          (checked) => this._toggleChanged('show_maintenance_summary', checked, false))}
-        <div class="description">${localize('editor.show_maintenance_summary_desc')}</div>
-
-        ${this._config.show_maintenance_summary === true ? html`
-          <div style="margin-left: 26px; margin-bottom: 8px;">
-            ${this._renderCheckbox('show-maintenance-activity', localize('editor.show_maintenance_activity'), this._config.show_maintenance_activity !== false,
-              (checked) => this._toggleChanged('show_maintenance_activity', checked, true))}
-            <div class="description">${localize('editor.show_maintenance_activity_desc')}</div>
-
-            ${this._renderCheckbox('show-video-tips', localize('editor.show_video_tips'), this._config.show_video_tips !== false,
-              (checked) => this._toggleChanged('show_video_tips', checked, true))}
-            <div class="description">${localize('editor.show_video_tips_desc')}</div>
-
-            ${this._renderMaintenanceUsersPicker()}
-          </div>
-        ` : nothing}
-      </div>
-    `;
-  }
-
-  /**
-   * User restriction for the maintenance tile/view. Options come from
-   * person entities carrying a user_id — no admin-only WS call needed.
-   * Empty selection state (no key in config) = visible to everyone.
-   */
-  private _renderMaintenanceUsersPicker(): TemplateResult {
-    if (!this._hass) return html``;
-    const options: { userId: string; name: string }[] = [];
-    for (const [entityId, state] of Object.entries(this._hass.states)) {
-      if (!entityId.startsWith('person.')) continue;
-      const userId = state.attributes?.user_id as string | undefined;
-      if (!userId) continue;
-      options.push({
-        userId,
-        name: (state.attributes?.friendly_name as string | undefined) || entityId,
-      });
-    }
-
-    const selected = this._config.maintenance_visible_users || [];
-    const allVisible = selected.length === 0;
-
-    return html`
-      <div style="font-size: 13px; font-weight: 500; color: var(--primary-text-color); margin-top: 4px; margin-bottom: 4px;">
-        ${localize('editor.maintenance_visible_users')}
-      </div>
-      <div class="description" style="margin-left: 0;">
-        ${localize('editor.maintenance_visible_users_desc')}
-      </div>
-      ${options.length === 0
-        ? html`<div class="description" style="margin-left: 0;">${localize('editor.maintenance_visible_users_none')}</div>`
-        : options.map((opt) => this._renderCheckbox(
-            `maintenance-user-${opt.userId}`,
-            opt.name,
-            allVisible || selected.includes(opt.userId),
-            (checked) => this._maintenanceUserChanged(opt.userId, options.map((o) => o.userId), checked)
-          ))}
-    `;
-  }
-
-  private _maintenanceUserChanged(userId: string, allUserIds: string[], checked: boolean): void {
-    const current = this._config.maintenance_visible_users || [];
-    // No restriction stored = everyone checked; start from the full set
-    const effective = new Set(current.length > 0 ? current : allUserIds);
-    if (checked) effective.add(userId);
-    else effective.delete(userId);
-
-    const updated: Simon42StrategyConfig = { ...this._config };
-    const known = allUserIds.filter((id) => effective.has(id));
-    // Ids configured via YAML for users without a person entity — keep them
-    const unknown = [...effective].filter((id) => !allUserIds.includes(id));
-    if (unknown.length === 0 && known.length === allUserIds.length) {
-      delete updated.maintenance_visible_users; // everyone = no restriction
-    } else {
-      updated.maintenance_visible_users = [...known, ...unknown];
-    }
-    this._fireConfigChanged(updated);
-  }
-
-  private _unavailableBatteriesBucketChanged(bucket: 'critical' | 'good'): void {
-    const updated: Simon42StrategyConfig = { ...this._config };
-    // 'good' is now the default → omit the key when matching default
-    if (bucket === 'good') {
-      delete updated.unavailable_batteries_bucket;
-    } else {
-      updated.unavailable_batteries_bucket = bucket;
-    }
-    this._fireConfigChanged(updated);
-  }
-
-  private _securityActivityPositionChanged(atEnd: boolean): void {
-    const updated: Simon42StrategyConfig = { ...this._config };
-    // 'start' is the default → omit the key when matching default
-    if (atEnd) {
-      updated.security_activity_position = 'end';
-    } else {
-      delete updated.security_activity_position;
-    }
-    this._config = updated;
-    this._fireConfigChanged(updated);
-  }
-
-  private _cameraHiddenChanged(entityId: string, visible: boolean): void {
-    const hidden = new Set(this._config.hidden_cameras || []);
-    if (visible) hidden.delete(entityId);
-    else hidden.add(entityId);
-    const updated: Simon42StrategyConfig = { ...this._config };
-    if (hidden.size === 0) {
-      delete updated.hidden_cameras;
-    } else {
-      updated.hidden_cameras = [...hidden].sort();
-    }
-    this._config = updated;
-    this._fireConfigChanged(updated);
-  }
-
-  /** Per-camera visibility for the security view (security-only exclusion). */
-  private _renderHiddenCamerasPicker(): TemplateResult {
-    if (!this._hass) return html``;
-    // Same dedup as the views (one camera per device, preferred stream);
-    // Registry is initialized by the dashboard render, this is a no-op.
-    Registry.initialize(this._hass, this._config);
-    const blocks = collectCameraBlocks(this._hass, this._config);
-    if (blocks.length === 0) return html``;
-
-    const hidden = new Set(this._config.hidden_cameras || []);
-    return html`
-      <div style="margin-left: 26px; margin-bottom: 8px;">
-        <div style="font-size: 13px; font-weight: 500; color: var(--primary-text-color); margin-top: 4px; margin-bottom: 4px;">
-          ${localize('editor.security_cameras_visibility')}
-        </div>
-        <div class="description" style="margin-left: 0;">
-          ${localize('editor.security_cameras_visibility_desc')}
-        </div>
-        ${blocks.map((block) => {
-          const name =
-            (this._hass?.states[block.cameraId]?.attributes?.friendly_name as string | undefined) ||
-            block.cameraId;
-          return this._renderCheckbox(
-            `security-camera-${block.cameraId}`,
-            name,
-            !hidden.has(block.cameraId),
-            (checked) => this._cameraHiddenChanged(block.cameraId, checked)
-          );
-        })}
-      </div>
-    `;
-  }
-
-  private _renderSecurityExtraEntitiesPicker(): TemplateResult {
-    const extras = this._config.security_extra_entities || [];
-    const allEntities = getAllEntitiesForSelect(this._hass);
-    const entityMap = new Map(allEntities.map((e) => [e.entity_id, e.name]));
-    const filtered = getFilteredEntities(this._hass, this._securityExtraSearch);
-    return html`
-      <div style="font-size: 13px; font-weight: 500; color: var(--primary-text-color); margin-top: 4px; margin-bottom: 4px;">
-        ${localize('editor.security_extra_entities')}
-      </div>
-      <div class="description" style="margin-left: 0; margin-bottom: 8px;">
-        ${localize('editor.security_extra_entities_desc')}
-      </div>
-      ${extras.length > 0 ? html`
-        <div class="entity-list-container" style="margin-bottom: 8px;">
-          ${extras.map((entityId) => {
-            const name = entityMap.get(entityId) || entityId;
-            return html`
-              <div class="entity-list-item" data-entity-id=${entityId}>
-                <span class="item-info">
-                  <span class="item-name">${name}</span>
-                  <span class="item-entity-id">${entityId}</span>
-                </span>
-                <button class="btn-remove" @click=${() => this._removeSecurityExtraEntity(entityId)}>&#x2715;</button>
-              </div>
-            `;
-          })}
-        </div>
-      ` : nothing}
-      <div class="entity-search-picker">
-        <input type="text" class="entity-search-input"
-          placeholder=${localize('editor.select_entity') + '...'}
-          .value=${this._securityExtraSearch}
-          @input=${(e: Event) => { this._securityExtraSearch = (e.target as HTMLInputElement).value; this.requestUpdate(); }}
-          @blur=${() => { setTimeout(() => { this._securityExtraSearch = ''; this.requestUpdate(); }, 200); }}
-        />
-        ${this._securityExtraSearch.length >= 2 ? html`
-          <div class="entity-search-results">
-            ${filtered.length > 0
-              ? filtered.map((entity) => html`
-                <div class="entity-search-result" @mousedown=${(e: Event) => { e.preventDefault(); this._addSecurityExtraEntity(entity.entity_id); this._securityExtraSearch = ''; this.requestUpdate(); }}>
-                  <span class="entity-search-name">${entity.name}</span>
-                  <span class="entity-search-id">${entity.entity_id}</span>
-                </div>
-              `)
-              : html`<div class="entity-search-no-results">${localize('editor.no_results')}</div>`
-            }
-          </div>
-        ` : nothing}
-      </div>
-    `;
-  }
-
-  private _addSecurityExtraEntity(entityId: string): void {
-    const current = this._config.security_extra_entities || [];
-    if (current.includes(entityId)) return;
-    const updated: Simon42StrategyConfig = { ...this._config, security_extra_entities: [...current, entityId] };
-    this._fireConfigChanged(updated);
-  }
-
-  private _removeSecurityExtraEntity(entityId: string): void {
-    const current = this._config.security_extra_entities || [];
-    const next = current.filter((e) => e !== entityId);
-    const updated: Simon42StrategyConfig = { ...this._config };
-    if (next.length === 0) {
-      delete updated.security_extra_entities;
-    } else {
-      updated.security_extra_entities = next;
-    }
-    this._fireConfigChanged(updated);
-  }
-
-  // -- Weather sensors editor -------------------------------------------
-  //
-  // Per-row structured editor for the `weather_sensors` config array.
-  // Each row binds to a WeatherSensorConfig and exposes inline inputs for
-  // icon / unit / round. Adding a row uses the same entity-search picker
-  // pattern as favorites; removal is a single-click button.
-  //
-  // The picker filters to numeric-ish sensors by default but does not hard-
-  // restrict — any entity domain is accepted (the markdown row in the
-  // section renderer just calls `states(...)` against the id).
 
   private _renderAreasSection(): TemplateResult {
     const groupByFloors = this._config.group_by_floors === true;
@@ -1388,19 +986,6 @@ class Simon42DashboardStrategyEditor extends LitElement implements StrategyEdito
   // EVENT HANDLERS — Toggle / Config changes
   // ====================================================================
 
-  /** lights_sort_by is an enum with a boolean-shaped editor toggle:
-   *  checked → 'name', unchecked → remove (default 'last_changed'). */
-  private _lightsSortByChanged(checked: boolean): void {
-    const newConfig: Simon42StrategyConfig = { ...this._config };
-    if (checked) {
-      newConfig.lights_sort_by = 'name';
-    } else {
-      delete newConfig.lights_sort_by;
-    }
-    this._config = newConfig;
-    this._fireConfigChanged(newConfig);
-  }
-
   _toggleChanged(key: string, value: boolean, defaultValue: boolean): void {
     if (!this._hass) return;
 
@@ -1414,40 +999,6 @@ class Simon42DashboardStrategyEditor extends LitElement implements StrategyEdito
       delete (newConfig as any)[key];
     }
 
-    this._config = newConfig;
-    this._fireConfigChanged(newConfig);
-  }
-
-  private _summariesColumnsChanged(columns: 2 | 4): void {
-    if (!this._hass) return;
-
-    const newConfig: Simon42StrategyConfig = {
-      ...this._config,
-      summaries_columns: columns,
-    };
-
-    if (columns === 2) {
-      delete newConfig.summaries_columns;
-    }
-
-    this._config = newConfig;
-    this._fireConfigChanged(newConfig);
-  }
-
-  private _batteryCriticalChanged(e: Event): void {
-    const value = parseInt((e.target as HTMLInputElement).value, 10);
-    if (isNaN(value) || value < 1 || value > 99) return;
-    const newConfig: Simon42StrategyConfig = { ...this._config, battery_critical_threshold: value };
-    if (value === 20) delete newConfig.battery_critical_threshold;
-    this._config = newConfig;
-    this._fireConfigChanged(newConfig);
-  }
-
-  private _batteryLowChanged(e: Event): void {
-    const value = parseInt((e.target as HTMLInputElement).value, 10);
-    if (isNaN(value) || value < 1 || value > 99) return;
-    const newConfig: Simon42StrategyConfig = { ...this._config, battery_low_threshold: value };
-    if (value === 50) delete newConfig.battery_low_threshold;
     this._config = newConfig;
     this._fireConfigChanged(newConfig);
   }
